refactor: separate files for renderer process
