+++ src/renderer/about/about.html
@@ -18,7 +18,11 @@
 					<h1>apm<br /><span class="lead">AviUtl Package Manager</span></h1>
 				</div>
 				<div class="col-3">
-					<img src="../icon/apm1024.png" alt="apm Logo" class="img-thumbnail" />
+					<img
+						src="../../../icon/apm1024.png"
+						alt="apm Logo"
+						class="img-thumbnail"
+					/>
 				</div>
 			</div>
