+++ src/renderer/main/index.html
@@ -65,7 +65,11 @@
 					<nav class="navbar navbar-light">
 						<div class="container-fluid">
 							<span class="navbar-brand">
-								<img src="../icon/apm32.png" alt="" class="d-inline-block" />
+								<img
+									src="../../../icon/apm32.png"
+									alt=""
+									class="d-inline-block"
+								/>
 								<span class="align-middle">AviUtl Package Manager</span>
 							</span>
 						</div>
