+++ src/renderer/main/renderer.ts
@@ -1,0 +1,5 @@
+import '../../../node_modules/bootstrap-dark-5/dist/css/bootstrap-dark.min.css';
+import '../../../node_modules/bootstrap-icons/font/bootstrap-icons.css';
+import '../main.css';
+import './index.css';
+import 'bootstrap/dist/js/bootstrap.bundle.min';
