+++ src/renderer/splash/splash.css
