+++ src/renderer/splash/splash.html
