--- src/splash_renderer.ts
@@ -1,4 +1,0 @@
-import '../node_modules/bootstrap/dist/css/bootstrap.min.css';
-import './main.css';
-import './splash.css';
-import 'bootstrap/dist/js/bootstrap.bundle.min';
